+++ src/plugins/views/editors/Flow/Flow.jsx
@@ -37,8 +37,6 @@ import { FLOW_VIEW_MODE } from "./Constants/constants";
 import GraphBase from "./Core/Graph/GraphBase";
 import GraphTreeView from "./Core/Graph/GraphTreeView";
 
-import * as d3 from "d3";
-
 import "./Resources/css/Flow.css";
 import { flowStyles } from "./styles";
 let activeBookmark = null;
@@ -1033,31 +1031,14 @@ const Flow = (props, ref) => {
    * Handle zoom reset
    */
   const handleResetZoom = useCallback(_e => {
-    const { canvas } = getMainInterface();
-    canvas
-      .getSvg()
-      .transition()
-      .duration(750)
-      .call(canvas.zoomBehavior.transform, d3.zoomIdentity);
+    getMainInterface()?.onResetZoom();
   }, []);
 
   /**
    * Handle Move Node
    */
-  const handleMoveNode = useCallback(event => {
-    const mainInterface = getMainInterface();
-    const currentZoom = mainInterface.canvas.currentZoom?.k ?? 1;
-    const step = 2 / currentZoom + 1;
-    const delta = {
-      ArrowRight: [1 * step, 0],
-      ArrowLeft: [-1 * step, 0],
-      ArrowUp: [0, -1 * step],
-      ArrowDown: [0, 1 * step]
-    };
-    const [dx, dy] = delta[event.code];
-    const [x, y] = [50, 50]; // skip boundaries validation used when dragging a node
-    mainInterface.graph.onNodeDrag(null, { x, y, dx, dy });
-    mainInterface.onDragEnd();
+  const handleMoveNode = useCallback(e => {
+    getMainInterface()?.onMoveNode(e);
   }, []);
 
   //========================================================================================
@@ -1105,6 +1086,8 @@ const Flow = (props, ref) => {
       KEYBINDINGS.FLOW.KEYBINDS.PASTE_NODE.SHORTCUTS,
       handlePasteNodes
     );
+    addKeyBind(KEYBINDINGS.FLOW.KEYBINDS.MOVE_NODE.SHORTCUTS, handleMoveNode);
+    addKeyBind(KEYBINDINGS.FLOW.KEYBINDS.RESET_ZOOM.SHORTCUTS, handleResetZoom);
     addKeyBind(
       KEYBINDINGS.EDITOR_GENERAL.KEYBINDS.CANCEL.SHORTCUTS,
       setFlowsToDefault
@@ -1117,6 +1100,8 @@ const Flow = (props, ref) => {
     return () => {
       removeKeyBind(KEYBINDINGS.FLOW.KEYBINDS.COPY_NODE.SHORTCUTS);
       removeKeyBind(KEYBINDINGS.FLOW.KEYBINDS.PASTE_NODE.SHORTCUTS);
+      removeKeyBind(KEYBINDINGS.FLOW.KEYBINDS.MOVE_NODE.SHORTCUTS);
+      removeKeyBind(KEYBINDINGS.FLOW.KEYBINDS.RESET_ZOOM.SHORTCUTS);
       removeKeyBind(KEYBINDINGS.EDITOR_GENERAL.KEYBINDS.CANCEL.SHORTCUTS);
       removeKeyBind(KEYBINDINGS.EDITOR_GENERAL.KEYBINDS.DELETE.SHORTCUTS);
     };