+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - [FP-1741 - Push flow parameter errors to Warnings, and make warnings clickable](https://movai.atlassian.net/browse/FP-1741)
 - [FP-1768 - Grey out other links when creating a new link](https://movai.atlassian.net/browse/FP-1768)
 - [FP-1794 - Add invalid links to warnings](https://movai.atlassian.net/browse/FP-1794)
+- [FP-1839 - New snackbar is crashing in all apps](https://movai.atlassian.net/browse/FP-1839)
+- [FP-1853 - Migrate Flow shortcuts - Reset Zoom and move nodes with arrow keys](https://movai.atlassian.net/browse/FP-1853)
 
 # 3.0.0